Clustering using Epigraph and Hypograph indices

It creates a multivariate dataset containing the epigraph, hypograph and/or its modified versions on the curves and derivatives and then perform hierarchical clustering, kmeans, kernel kmeans, and spectral clustering

Usage

EHyClus(
  curves,
  vars_combinations,
  k = 30,
  n_clusters = 2,
  bs = "cr",
  clustering_methods = c("hierarch", "kmeans", "kkmeans", "spc"),
  l_method_hierarch = c("single", "complete", "average", "centroid", "ward.D2"),
  l_dist_hierarch = c("euclidean", "manhattan"),
  l_dist_kmeans = c("euclidean", "mahalanobis"),
  l_kernel = c("rbfdot", "polydot"),
  grid,
  true_labels = NULL,
  only_best = FALSE,
  verbose = FALSE,
  n_cores = 1
)

Arguments

curves

Dataset containing the curves to apply a clustering algorithm. The functional dataset can be one dimensional (\(n \times p\)) where \(n\) is the number of curves and p the number of time points, or multidimensional (\(n \times p \times q\)) where \(q\) represents the number of dimensions in the data

vars_combinations

If list, each element of the list should be an atomic vector of strings with the names of the variables. Combinations with non-valid variable names will be discarded. If the list is non-named, the names of the variables are set to vars1, ..., varsk, where k is the number of elements in vars_combinations. If not provided, generic combinations of variables will be used. They will not be the same for uni-dimensional and multi-dimensional problems.

k

Number of basis functions for the B-splines. If equals to 0, the number of basis functions will be automatically selected.

n_clusters

Number of clusters to generate.

bs

A two letter character string indicating the (penalized) smoothing basis to use. See smooth.terms.

clustering_methods

character vector specifying at least one of the following clustering methods to be computed: "hierarch", "kmeans", "kkmeans" or "spc".

l_method_hierarch

list of clustering methods for hierarchical clustering.

l_dist_hierarch

list of distances for hierarchical clustering.

l_dist_kmeans

list of distances for kmeans clustering.

l_kernel

list of kernels for kkmeans or spc.

grid

Atomic vector of type numeric with two elements: the lower limit and the upper limit of the evaluation grid. If not provided, it will be selected automatically.

true_labels

Numeric vector of true labels for validation. If provided, evaluation metrics are computed in the final result.

only_best

logical value. If TRUE and true_labels is provided, the function will return only the result for the best clustering method based on the Rand Index. Defaults to FALSE.

verbose

If TRUE, the function will print logs for about the execution of some clustering methods. Defaults to FALSE.

n_cores

Number of cores to do parallel computation. 1 by default, which mean no parallel execution. Must be an integer number greater than 1.

Value

A list containing the clustering partition for each method and indices combination and, if true_labels is provided a data frame containing the time elapsed for obtaining a clustering partition of the indices dataset for each methodology. Also, the number of generated clusters and the combinations of variables used can be seen as attributes of this object.

Examples

# univarariate data without labels
curves <- sim_model_ex1(n = 10)
vars_combinations <- list(c("dtaEI", "dtaMEI"), c("dtaHI", "dtaMHI"))
EHyClus(curves, vars_combinations = vars_combinations)
#> $cluster
#> $cluster$hierarch
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaMEI$time
#> [1] 0.0004923344
#> 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_single_euclidean_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaHIdtaMHI$time
#> [1] 0.0001795292
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaMEI$time
#> [1] 0.0001716614
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_complete_euclidean_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaHIdtaMHI$time
#> [1] 0.0001673698
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaMEI$time
#> [1] 0.0001633167
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_average_euclidean_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaHIdtaMHI$time
#> [1] 0.0001635551
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaMEI$time
#> [1] 0.0001637936
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaHIdtaMHI$time
#> [1] 0.0001528263
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  1  1  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaMEI$time
#> [1] 0.0001616478
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaHIdtaMHI$time
#> [1] 0.0001542568
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaMEI$time
#> [1] 0.000150919
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_single_manhattan_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaHIdtaMHI$time
#> [1] 0.0001559258
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaMEI$time
#> [1] 0.0001513958
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_complete_manhattan_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaHIdtaMHI$time
#> [1] 0.0001525879
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaMEI$time
#> [1] 0.0001487732
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_average_manhattan_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaHIdtaMHI$time
#> [1] 0.0001502037
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaMEI$time
#> [1] 0.0001683235
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaHIdtaMHI$time
#> [1] 0.0001544952
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaMEI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaMEI$time
#> [1] 0.0001487732
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaHIdtaMHI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  1  1  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaHIdtaMHI$time
#> [1] 0.0001535416
#> 
#> 
#> 
#> $cluster$kmeans
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaMEI
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  2  1  2  1  1  1  2  1  1  1  2  2  2  2  1  1  2  2  2  2 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaMEI$time
#> [1] 0.002938271
#> 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaHIdtaMHI
#> $cluster$kmeans$kmeans_euclidean_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  1  2  1  2  2  2  1  2  2  2  1  1  1  1  2  2  1  1  1  1 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaHIdtaMHI$time
#> [1] 0.002206802
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaMEI
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  2  1  2  1  1  1  2  1  1  1  2  2  2  2  1  1  2  2  2  2 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaMEI$time
#> [1] 0.003036499
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaHIdtaMHI
#> $cluster$kmeans$kmeans_mahalanobis_dtaHIdtaMHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 
#>  2  2  2  2  2  2  2  2  2  2  1  2  1  2  2  2  1  2  2  2 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaHIdtaMHI$time
#> [1] 0.002320528
#> 
#> 
#> 
#> $cluster$kkmeans
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaMEI
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaMEI$cluster
#>  [1] 2 1 2 1 1 1 2 1 1 1 2 2 2 2 1 1 2 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaMEI$time
#> [1] 0.4877985
#> 
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaHIdtaMHI
#> $cluster$kkmeans$kkmeans_rbfdot_dtaHIdtaMHI$cluster
#>  [1] 1 2 1 2 2 2 1 2 2 2 1 2 1 2 1 1 1 1 1 1
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaHIdtaMHI$time
#> [1] 0.1417618
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaMEI
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaMEI$cluster
#>  [1] 2 1 2 1 1 1 2 1 1 1 2 2 2 2 1 1 2 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaMEI$time
#> [1] 0.006959438
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaHIdtaMHI
#> $cluster$kkmeans$kkmeans_polydot_dtaHIdtaMHI$cluster
#>  [1] 2 1 2 1 1 1 2 1 1 1 2 2 2 2 1 1 2 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaHIdtaMHI$time
#> [1] 0.005225182
#> 
#> 
#> 
#> $cluster$spc
#> $cluster$spc$spc_rbfdot_dtaEIdtaMEI
#> $cluster$spc$spc_rbfdot_dtaEIdtaMEI$cluster
#>  [1] 1 2 1 1 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> 
#> $cluster$spc$spc_rbfdot_dtaEIdtaMEI$time
#> [1] 0.02830577
#> 
#> 
#> $cluster$spc$spc_rbfdot_dtaHIdtaMHI
#> $cluster$spc$spc_rbfdot_dtaHIdtaMHI$cluster
#>  [1] 2 2 2 2 2 2 2 2 2 2 1 2 1 2 2 2 1 2 2 2
#> 
#> $cluster$spc$spc_rbfdot_dtaHIdtaMHI$time
#> [1] 0.1502569
#> 
#> 
#> $cluster$spc$spc_polydot_dtaEIdtaMEI
#> $cluster$spc$spc_polydot_dtaEIdtaMEI$cluster
#>  [1] 2 1 2 2 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> 
#> $cluster$spc$spc_polydot_dtaEIdtaMEI$time
#> [1] 0.01540804
#> 
#> 
#> $cluster$spc$spc_polydot_dtaHIdtaMHI
#> $cluster$spc$spc_polydot_dtaHIdtaMHI$cluster
#>  [1] 1 1 1 1 1 1 1 1 1 1 2 1 2 1 1 1 2 1 1 1
#> 
#> $cluster$spc$spc_polydot_dtaHIdtaMHI$time
#> [1] 0.01556516
#> 
#> 
#> 
#> 
#> attr(,"class")
#> [1] "EHyClus" "list"   
#> attr(,"n_clusters")
#> [1] 2
#> attr(,"vars_combinations")
#> attr(,"vars_combinations")[[1]]
#> [1] "dtaEI"  "dtaMEI"
#> 
#> attr(,"vars_combinations")[[2]]
#> [1] "dtaHI"  "dtaMHI"
#> 

# multivariate data with labels
curves <- sim_model_ex2(n = 5)
true_labels <- c(rep(1, 5), rep(2, 5))
vars_combinations <- list(c("dtaMEI", "ddtaMEI"), c("dtaMEI", "d2dtaMEI"))
res <- EHyClus(curves, vars_combinations = vars_combinations, true_labels = true_labels)
res$cluster # clustering results
#> $hierarch
#> $hierarch$hierarch_single_euclidean_dtaMEIddtaMEI
#> $hierarch$hierarch_single_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_single_euclidean_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_single_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0002758503
#> 
#> 
#> $hierarch$hierarch_single_euclidean_dtaMEId2dtaMEI
#> $hierarch$hierarch_single_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_single_euclidean_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_single_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0002291203
#> 
#> 
#> $hierarch$hierarch_complete_euclidean_dtaMEIddtaMEI
#> $hierarch$hierarch_complete_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_complete_euclidean_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_complete_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0002167225
#> 
#> 
#> $hierarch$hierarch_complete_euclidean_dtaMEId2dtaMEI
#> $hierarch$hierarch_complete_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_complete_euclidean_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_complete_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0002379417
#> 
#> 
#> $hierarch$hierarch_average_euclidean_dtaMEIddtaMEI
#> $hierarch$hierarch_average_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_average_euclidean_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_average_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001590252
#> 
#> 
#> $hierarch$hierarch_average_euclidean_dtaMEId2dtaMEI
#> $hierarch$hierarch_average_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_average_euclidean_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_average_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0001487732
#> 
#> 
#> $hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEI
#> $hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001497269
#> 
#> 
#> $hierarch$hierarch_centroid_euclidean_dtaMEId2dtaMEI
#> $hierarch$hierarch_centroid_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_centroid_euclidean_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_centroid_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0001523495
#> 
#> 
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEI
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001597404
#> 
#> 
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEId2dtaMEI
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_ward.D2_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0001511574
#> 
#> 
#> $hierarch$hierarch_single_manhattan_dtaMEIddtaMEI
#> $hierarch$hierarch_single_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_single_manhattan_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_single_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001533031
#> 
#> 
#> $hierarch$hierarch_single_manhattan_dtaMEId2dtaMEI
#> $hierarch$hierarch_single_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_single_manhattan_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_single_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001525879
#> 
#> 
#> $hierarch$hierarch_complete_manhattan_dtaMEIddtaMEI
#> $hierarch$hierarch_complete_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_complete_manhattan_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_complete_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001451969
#> 
#> 
#> $hierarch$hierarch_complete_manhattan_dtaMEId2dtaMEI
#> $hierarch$hierarch_complete_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_complete_manhattan_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_complete_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001554489
#> 
#> 
#> $hierarch$hierarch_average_manhattan_dtaMEIddtaMEI
#> $hierarch$hierarch_average_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_average_manhattan_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_average_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001468658
#> 
#> 
#> $hierarch$hierarch_average_manhattan_dtaMEId2dtaMEI
#> $hierarch$hierarch_average_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_average_manhattan_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_average_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001590252
#> 
#> 
#> $hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEI
#> $hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001468658
#> 
#> 
#> $hierarch$hierarch_centroid_manhattan_dtaMEId2dtaMEI
#> $hierarch$hierarch_centroid_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_centroid_manhattan_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_centroid_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001528263
#> 
#> 
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEI
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001525879
#> 
#> 
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEId2dtaMEI
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $hierarch$hierarch_ward.D2_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001549721
#> 
#> 
#> 
#> $kmeans
#> $kmeans$kmeans_euclidean_dtaMEIddtaMEI
#> $kmeans$kmeans_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $kmeans$kmeans_euclidean_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $kmeans$kmeans_euclidean_dtaMEIddtaMEI$time
#> [1] 0.002351761
#> 
#> 
#> $kmeans$kmeans_euclidean_dtaMEId2dtaMEI
#> $kmeans$kmeans_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $kmeans$kmeans_euclidean_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $kmeans$kmeans_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.00225544
#> 
#> 
#> $kmeans$kmeans_mahalanobis_dtaMEIddtaMEI
#> $kmeans$kmeans_mahalanobis_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  1  1  2  1  1  2  2  2  2 
#> 
#> $kmeans$kmeans_mahalanobis_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $kmeans$kmeans_mahalanobis_dtaMEIddtaMEI$time
#> [1] 0.002392292
#> 
#> 
#> $kmeans$kmeans_mahalanobis_dtaMEId2dtaMEI
#> $kmeans$kmeans_mahalanobis_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  2  2  1  2  2  1  1  1  1 
#> 
#> $kmeans$kmeans_mahalanobis_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $kmeans$kmeans_mahalanobis_dtaMEId2dtaMEI$time
#> [1] 0.002357483
#> 
#> 
#> 
#> $kkmeans
#> $kkmeans$kkmeans_rbfdot_dtaMEIddtaMEI
#> $kkmeans$kkmeans_rbfdot_dtaMEIddtaMEI$cluster
#>  [1] 1 2 2 1 1 1 1 1 1 1
#> 
#> $kkmeans$kkmeans_rbfdot_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.5714   0.5333 
#> 
#> $kkmeans$kkmeans_rbfdot_dtaMEIddtaMEI$time
#> [1] 0.004669428
#> 
#> 
#> $kkmeans$kkmeans_rbfdot_dtaMEId2dtaMEI
#> $kkmeans$kkmeans_rbfdot_dtaMEId2dtaMEI$cluster
#>  [1] 1 2 2 2 1 1 1 1 1 2
#> 
#> $kkmeans$kkmeans_rbfdot_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $kkmeans$kkmeans_rbfdot_dtaMEId2dtaMEI$time
#> [1] 0.004255295
#> 
#> 
#> $kkmeans$kkmeans_polydot_dtaMEIddtaMEI
#> $kkmeans$kkmeans_polydot_dtaMEIddtaMEI$cluster
#>  [1] 1 2 2 1 2 2 1 1 1 1
#> 
#> $kkmeans$kkmeans_polydot_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $kkmeans$kkmeans_polydot_dtaMEIddtaMEI$time
#> [1] 0.003573179
#> 
#> 
#> $kkmeans$kkmeans_polydot_dtaMEId2dtaMEI
#> $kkmeans$kkmeans_polydot_dtaMEId2dtaMEI$cluster
#>  [1] 1 2 2 1 2 2 1 1 1 1
#> 
#> $kkmeans$kkmeans_polydot_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $kkmeans$kkmeans_polydot_dtaMEId2dtaMEI$time
#> [1] 0.003459215
#> 
#> 
#> 
#> $spc
#> $spc$spc_rbfdot_dtaMEIddtaMEI
#> $spc$spc_rbfdot_dtaMEIddtaMEI$cluster
#>  [1] 2 1 1 2 1 1 2 2 2 2
#> 
#> $spc$spc_rbfdot_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $spc$spc_rbfdot_dtaMEIddtaMEI$time
#> [1] 0.00963378
#> 
#> 
#> $spc$spc_rbfdot_dtaMEId2dtaMEI
#> $spc$spc_rbfdot_dtaMEId2dtaMEI$cluster
#>  [1] 1 2 2 1 2 2 1 1 1 1
#> 
#> $spc$spc_rbfdot_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $spc$spc_rbfdot_dtaMEId2dtaMEI$time
#> [1] 0.009691
#> 
#> 
#> $spc$spc_polydot_dtaMEIddtaMEI
#> $spc$spc_polydot_dtaMEIddtaMEI$cluster
#>  [1] 2 1 1 2 1 1 2 2 2 2
#> 
#> $spc$spc_polydot_dtaMEIddtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $spc$spc_polydot_dtaMEIddtaMEI$time
#> [1] 0.009058714
#> 
#> 
#> $spc$spc_polydot_dtaMEId2dtaMEI
#> $spc$spc_polydot_dtaMEId2dtaMEI$cluster
#>  [1] 1 2 2 1 2 2 1 1 1 1
#> 
#> $spc$spc_polydot_dtaMEId2dtaMEI$valid
#>   Purity Fmeasure       RI 
#>   0.7000   0.4878   0.5333 
#> 
#> $spc$spc_polydot_dtaMEId2dtaMEI$time
#> [1] 0.01029611
#> 
#> 
#> 

# multivariate data and generic (default) vars_combinations
curves <- sim_model_ex2(n = 5)
EHyClus(curves)
#> Warning: Combination/s of variables with index/indices 2, 3, 4, 5, 6, 7, 13, 14, 15, 16, 17, 18 is/are singular or almost singular. Removing them...
#> Warning: Combination/s of variable/s with index 2, 3, 4, 5, 6, 7, 13, 14, 15, 16, 17, 18 are not valid. Excluding them from any computation...
#> $cluster
#> $cluster$hierarch
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaHI
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaHI$time
#> [1] 0.0002074242
#> 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001516342
#> 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.000150919
#> 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_single_euclidean_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  2  1 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_ddtaMEId2dtaMEI$time
#> [1] 0.0001432896
#> 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001425743
#> 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_single_euclidean_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001442432
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaHI
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaHI$time
#> [1] 0.0001351833
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001432896
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0001358986
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_complete_euclidean_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  2  2  1  2  1 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_ddtaMEId2dtaMEI$time
#> [1] 0.0001497269
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001440048
#> 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_complete_euclidean_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001380444
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaHI
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaHI$time
#> [1] 0.0001392365
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001337528
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0001354218
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_average_euclidean_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_ddtaMEId2dtaMEI$time
#> [1] 0.0001344681
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001413822
#> 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_euclidean_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001392365
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaHI
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaHI$time
#> [1] 0.0001373291
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001480579
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.000161171
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_centroid_euclidean_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_ddtaMEId2dtaMEI$time
#> [1] 0.0002140999
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001468658
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_euclidean_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001454353
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaHI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaHI$time
#> [1] 0.0001409054
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEI$time
#> [1] 0.0001351833
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.0001423359
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  2  2  2  1  2  1 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_ddtaMEId2dtaMEI$time
#> [1] 0.0001401901
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001423359
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_euclidean_dtaEIdtaHIdtaMEI$time
#> [1] 0.000136137
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaHI
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaHI$time
#> [1] 0.0001301765
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001378059
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001409054
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_single_manhattan_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  2  1 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_ddtaMEId2dtaMEI$time
#> [1] 0.0001473427
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001485348
#> 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_single_manhattan_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001423359
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaHI
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaHI$time
#> [1] 0.0001449585
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001363754
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001444817
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_complete_manhattan_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  2  2  2  1  2  1 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_ddtaMEId2dtaMEI$time
#> [1] 0.0001337528
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  2  2  1  2 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001392365
#> 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_complete_manhattan_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001413822
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaHI
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaHI$time
#> [1] 0.0001399517
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001373291
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001363754
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_average_manhattan_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_ddtaMEId2dtaMEI$time
#> [1] 0.0001387596
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001773834
#> 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_average_manhattan_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001635551
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaHI
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaHI$time
#> [1] 0.0001437664
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001363754
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001370907
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_centroid_manhattan_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  2  1  1  1 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_ddtaMEId2dtaMEI$time
#> [1] 0.0001342297
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001382828
#> 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  1  1  2 
#> 
#> $cluster$hierarch$hierarch_centroid_manhattan_dtaEIdtaHIdtaMEI$time
#> [1] 0.000146389
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaHI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaHI$time
#> [1] 0.0001380444
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEI$time
#> [1] 0.0001392365
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEId2dtaMEI$time
#> [1] 0.0001347065
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_ddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  2  2  2  1  2  1 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_ddtaMEId2dtaMEI$time
#> [1] 0.0001399517
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEId2dtaMEI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.0001683235
#> 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaHIdtaMEI
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$hierarch$hierarch_ward.D2_manhattan_dtaEIdtaHIdtaMEI$time
#> [1] 0.0001440048
#> 
#> 
#> 
#> $cluster$kmeans
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaHI
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  2  2  2  2  2  2  1  2  1 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaHI$time
#> [1] 0.00232482
#> 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaMEIddtaMEI
#> $cluster$kmeans$kmeans_euclidean_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaMEIddtaMEI$time
#> [1] 0.002260923
#> 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaMEId2dtaMEI
#> $cluster$kmeans$kmeans_euclidean_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  2  2  1  2  2  1  1  1  1 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaMEId2dtaMEI$time
#> [1] 0.002209902
#> 
#> 
#> $cluster$kmeans$kmeans_euclidean_ddtaMEId2dtaMEI
#> $cluster$kmeans$kmeans_euclidean_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  2  2  1  2  1  1  2  1  2 
#> 
#> $cluster$kmeans$kmeans_euclidean_ddtaMEId2dtaMEI$time
#> [1] 0.002177
#> 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaMEIddtaMEId2dtaMEI
#> $cluster$kmeans$kmeans_euclidean_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.002202988
#> 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaHIdtaMEI
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  2  1  1  2  2  2  2 
#> 
#> $cluster$kmeans$kmeans_euclidean_dtaEIdtaHIdtaMEI$time
#> [1] 0.002190828
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaHI
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaHI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  1  1  2  2  2  2  2  2  2 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaHI$time
#> [1] 0.002215147
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEIddtaMEI
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEIddtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  1  1  1  1  1  1  2  2  1  2 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEIddtaMEI$time
#> [1] 0.002221107
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEId2dtaMEI
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  1  1  1  1  1  1  2  1  2 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEId2dtaMEI$time
#> [1] 0.002181768
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_ddtaMEId2dtaMEI
#> $cluster$kmeans$kmeans_mahalanobis_ddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  2  2  1  2  1  1  2  1  2 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_ddtaMEId2dtaMEI$time
#> [1] 0.002199888
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEIddtaMEId2dtaMEI
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEIddtaMEId2dtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  2  2  1  2  2  1  2  1  1 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.002631903
#> 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaHIdtaMEI
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaHIdtaMEI$cluster
#>  1  2  3  4  5  6  7  8  9 10 
#>  2  1  1  2  2  2  2  2  2  2 
#> 
#> $cluster$kmeans$kmeans_mahalanobis_dtaEIdtaHIdtaMEI$time
#> [1] 0.002205849
#> 
#> 
#> 
#> $cluster$kkmeans
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaHI
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaHI$cluster
#>  [1] 1 2 2 1 1 1 1 2 1 2
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaHI$time
#> [1] 0.004783154
#> 
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEIddtaMEI
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEIddtaMEI$cluster
#>  [1] 1 1 1 2 1 1 2 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEIddtaMEI$time
#> [1] 0.005424976
#> 
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEId2dtaMEI
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEId2dtaMEI$cluster
#>  [1] 1 1 1 2 1 1 2 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEId2dtaMEI$time
#> [1] 0.004154921
#> 
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_ddtaMEId2dtaMEI
#> $cluster$kkmeans$kkmeans_rbfdot_ddtaMEId2dtaMEI$cluster
#>  [1] 2 2 2 1 1 1 1 1 2 2
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_ddtaMEId2dtaMEI$time
#> [1] 0.002851248
#> 
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEIddtaMEId2dtaMEI
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEIddtaMEId2dtaMEI$cluster
#>  [1] 2 2 2 1 2 2 1 1 1 1
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.004115582
#> 
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaHIdtaMEI
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaHIdtaMEI$cluster
#>  [1] 1 2 1 2 1 1 1 1 2 2
#> 
#> $cluster$kkmeans$kkmeans_rbfdot_dtaEIdtaHIdtaMEI$time
#> [1] 0.004101276
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaHI
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaHI$cluster
#>  [1] 1 1 1 1 1 1 1 2 1 2
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaHI$time
#> [1] 0.003403187
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaMEIddtaMEI
#> $cluster$kkmeans$kkmeans_polydot_dtaMEIddtaMEI$cluster
#>  [1] 2 2 2 1 2 2 1 1 1 1
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaMEIddtaMEI$time
#> [1] 0.003400326
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaMEId2dtaMEI
#> $cluster$kkmeans$kkmeans_polydot_dtaMEId2dtaMEI$cluster
#>  [1] 1 1 1 2 1 1 2 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaMEId2dtaMEI$time
#> [1] 0.003377914
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_ddtaMEId2dtaMEI
#> $cluster$kkmeans$kkmeans_polydot_ddtaMEId2dtaMEI$cluster
#>  [1] 2 2 2 1 1 1 1 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_polydot_ddtaMEId2dtaMEI$time
#> [1] 0.003396034
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaMEIddtaMEId2dtaMEI
#> $cluster$kkmeans$kkmeans_polydot_dtaMEIddtaMEId2dtaMEI$cluster
#>  [1] 1 1 1 2 1 1 2 2 2 2
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaMEIddtaMEId2dtaMEI$time
#> [1] 0.003427267
#> 
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaHIdtaMEI
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaHIdtaMEI$cluster
#>  [1] 2 2 2 1 2 2 1 1 1 1
#> 
#> $cluster$kkmeans$kkmeans_polydot_dtaEIdtaHIdtaMEI$time
#> [1] 0.003554106
#> 
#> 
#> 
#> 
#> attr(,"class")
#> [1] "EHyClus" "list"   
#> attr(,"n_clusters")
#> [1] 2
#> attr(,"vars_combinations")
#> attr(,"vars_combinations")[[1]]
#> [1] "dtaEI" "dtaHI"
#> 
#> attr(,"vars_combinations")[[2]]
#> [1] "dtaMEI"  "ddtaMEI"
#> 
#> attr(,"vars_combinations")[[3]]
#> [1] "dtaMEI"   "d2dtaMEI"
#> 
#> attr(,"vars_combinations")[[4]]
#> [1] "ddtaMEI"  "d2dtaMEI"
#> 
#> attr(,"vars_combinations")[[5]]
#> [1] "dtaMEI"   "ddtaMEI"  "d2dtaMEI"
#> 
#> attr(,"vars_combinations")[[6]]
#> [1] "dtaEI"  "dtaHI"  "dtaMEI"
#>