Hypograph Index (HI) for a functional dataset

The Hypograph Index of a curve x is the proportion of curves in the sample that are below x.

Usage

HI(curves, ...)

Arguments

curves

matrix where each row represents a curve, and each column represents values along the curve or array with dimension \(n \times p \times q\) with \(n\) curves, \(p\) values along the curve, and \(q\) dimensions.

...

Ignored.

Value

numeric vector containing the HI for each curve.

Examples

x <- matrix(c(1, 2, 3, 3, 2, 1, 5, 2, 3, 9, 8, 7), ncol = 3, nrow = 4)
HI(x)
#> [1] 0.25 0.25 0.75 0.50

y <- array(c(1, 2, 3, 3, 2, 1, 5, 2, 3, 9, 8, 7, -1, -5, -6, 2, 3, 0, -1, 0, 2, -1, -2, 0),
  dim = c(3, 4, 2)
)
HI(y)
#> [1] 0.3333333 0.3333333 0.3333333