Perform hierarchical clustering for a different combinations of indices, method and distance

Perform hierarchical clustering for a different combinations of indices, method and distance

Usage

clustInd_hierarch(
  ind_data,
  vars_combinations,
  method_list = c("single", "complete", "average", "centroid", "ward.D2"),
  dist_vector = c("euclidean", "manhattan"),
  n_cluster = 2,
  true_labels = NULL,
  n_cores = 1
)

Arguments

ind_data

Dataframe containing indices applied to the original data and its first and second derivatives. See generate_indices.

vars_combinations

list containing one or more combinations of indices in ind_data. If it is non-named, the names of the variables are set to vars1, ..., varsk, where k is the number of elements in vars_combinations.

method_list

list of clustering methods.

dist_vector

list of distance metrics.

n_cluster

number of clusters to generate.

true_labels

Vector of true labels for validation (if it is not known true_labels is set to NULL)

n_cores

Number of cores to do parallel computation. 1 by default, which mean no parallel execution.

Value

A list containing hierarchical clustering results for each configuration.

Examples

vars1 <- c("dtaEI", "dtaMEI")
vars2 <- c("dtaHI", "dtaMHI")
data <- ehymet::sim_model_ex1()
data_ind <- generate_indices(data)
clustInd_hierarch(data_ind, list(vars1, vars2))
#> $hierarch_single_euclidean_dtaEIdtaMEI
#> $hierarch_single_euclidean_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   2   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   1   2   1   1   1   1   1   1   1   1   1   1   1 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#> 
#> $hierarch_single_euclidean_dtaEIdtaMEI$internal_metrics
#> $hierarch_single_euclidean_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.3540507
#> 
#> $hierarch_single_euclidean_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.1085543
#> 
#> $hierarch_single_euclidean_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.7129516
#> 
#> $hierarch_single_euclidean_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.1414405
#> 
#> 
#> $hierarch_single_euclidean_dtaEIdtaMEI$time
#> [1] 0.0004098415
#> 
#> 
#> $hierarch_single_euclidean_dtaHIdtaMHI
#> $hierarch_single_euclidean_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   1   2   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#> 
#> $hierarch_single_euclidean_dtaHIdtaMHI$internal_metrics
#> $hierarch_single_euclidean_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.4286669
#> 
#> $hierarch_single_euclidean_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.108974
#> 
#> $hierarch_single_euclidean_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.6451731
#> 
#> $hierarch_single_euclidean_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.1414405
#> 
#> 
#> $hierarch_single_euclidean_dtaHIdtaMHI$time
#> [1] 0.0003767014
#> 
#> 
#> $hierarch_complete_euclidean_dtaEIdtaMEI
#> $hierarch_complete_euclidean_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   2   2   2   1   1   2   2   2   2   2   2   1   2   2   1   2   2   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   2   2   2   2   1   2   1   2   2   2   2   2   1   1   2   1   2   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   2   2   2   2   1   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#> 
#> $hierarch_complete_euclidean_dtaEIdtaMEI$internal_metrics
#> $hierarch_complete_euclidean_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.5251489
#> 
#> $hierarch_complete_euclidean_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.0965386
#> 
#> $hierarch_complete_euclidean_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.611964
#> 
#> $hierarch_complete_euclidean_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.6343096
#> 
#> 
#> $hierarch_complete_euclidean_dtaEIdtaMEI$time
#> [1] 0.0003912449
#> 
#> 
#> $hierarch_complete_euclidean_dtaHIdtaMHI
#> $hierarch_complete_euclidean_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   1   1   1   1   1   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   2   2   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   2   1   1   1   2   2   2   2   1   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   1   2   2   1   1   2   1   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   1   1   2   1   2   1   1   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_complete_euclidean_dtaHIdtaMHI$internal_metrics
#> $hierarch_complete_euclidean_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.6103852
#> 
#> $hierarch_complete_euclidean_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.08346508
#> 
#> $hierarch_complete_euclidean_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.5562613
#> 
#> $hierarch_complete_euclidean_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.9765005
#> 
#> 
#> $hierarch_complete_euclidean_dtaHIdtaMHI$time
#> [1] 0.0004048347
#> 
#> 
#> $hierarch_average_euclidean_dtaEIdtaMEI
#> $hierarch_average_euclidean_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   2   1   1   1   1   1   1   2   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   2   1   2   1   1   1   1   1   1   2   1   2   1   2   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   2   1   1   1   2   1   1   1   1   1   1   1   1   1   1   1 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#> 
#> $hierarch_average_euclidean_dtaEIdtaMEI$internal_metrics
#> $hierarch_average_euclidean_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.4051792
#> 
#> $hierarch_average_euclidean_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.1077677
#> 
#> $hierarch_average_euclidean_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.6868052
#> 
#> $hierarch_average_euclidean_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.4364698
#> 
#> 
#> $hierarch_average_euclidean_dtaEIdtaMEI$time
#> [1] 0.0004327297
#> 
#> 
#> $hierarch_average_euclidean_dtaHIdtaMHI
#> $hierarch_average_euclidean_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   1   1   1   1   1   1   2   1   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   1   1   1   1   2   2   2   2   1   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   2   2   1   2   2   1   1   2   1   1   1   1   1   2   1   2   1   1   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   1   1   2   1   1   1   1   1   2   2   2   1   1   1   2   1   2 
#> 
#> $hierarch_average_euclidean_dtaHIdtaMHI$internal_metrics
#> $hierarch_average_euclidean_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.5564304
#> 
#> $hierarch_average_euclidean_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.06736565
#> 
#> $hierarch_average_euclidean_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.578379
#> 
#> $hierarch_average_euclidean_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.8112781
#> 
#> 
#> $hierarch_average_euclidean_dtaHIdtaMHI$time
#> [1] 0.0004210472
#> 
#> 
#> $hierarch_centroid_euclidean_dtaEIdtaMEI
#> $hierarch_centroid_euclidean_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   2   2   2   1   1   2   2   2   2   2   2   1   2   2   2   2   2   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   2   2   2   2   1   2   1   2   2   2   2   2   1   1   2   1   2   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   2   2   2   2   1   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#> 
#> $hierarch_centroid_euclidean_dtaEIdtaMEI$internal_metrics
#> $hierarch_centroid_euclidean_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.5057735
#> 
#> $hierarch_centroid_euclidean_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.07548191
#> 
#> $hierarch_centroid_euclidean_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.6235726
#> 
#> $hierarch_centroid_euclidean_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.6098403
#> 
#> 
#> $hierarch_centroid_euclidean_dtaEIdtaMEI$time
#> [1] 0.000415802
#> 
#> 
#> $hierarch_centroid_euclidean_dtaHIdtaMHI
#> $hierarch_centroid_euclidean_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   1   1   1   1   2   2   1   2   1   1   1   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   2   2   1   1   2   1   1   2   1   1   1   1   1   2   1   2   1   1   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   1   1   2   1   1   1   1   1   2   1   2   1   1   1   2   1   2 
#> 
#> $hierarch_centroid_euclidean_dtaHIdtaMHI$internal_metrics
#> $hierarch_centroid_euclidean_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.5274109
#> 
#> $hierarch_centroid_euclidean_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.06584967
#> 
#> $hierarch_centroid_euclidean_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.5985458
#> 
#> $hierarch_centroid_euclidean_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.7219281
#> 
#> 
#> $hierarch_centroid_euclidean_dtaHIdtaMHI$time
#> [1] 0.0004160404
#> 
#> 
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   2   1   1   1   2   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   2   1   2   1   1   2   2   2   1   1   1   2   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   2   2   2   1   1   1   2   1   2   2   2   2   2   2   2   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   1   2   1   2   1   2   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI$internal_metrics
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.6482362
#> 
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.05440341
#> 
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.5447726
#> 
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.9765005
#> 
#> 
#> $hierarch_ward.D2_euclidean_dtaEIdtaMEI$time
#> [1] 0.0004155636
#> 
#> 
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   2   1   1   1   2   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   2   1   1   1   1   2   2   2   1   1   1   2   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   2   1   2   1   1   1   2   1   2   2   2   2   2   2   2   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   1   2   1   2   1   2   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI$internal_metrics
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.6269101
#> 
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.03020551
#> 
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.5437642
#> 
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.985815
#> 
#> 
#> $hierarch_ward.D2_euclidean_dtaHIdtaMHI$time
#> [1] 0.0004234314
#> 
#> 
#> $hierarch_single_manhattan_dtaEIdtaMEI
#> $hierarch_single_manhattan_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   2   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   1   2   1   1   1   1   1   1   1   1   1   1   1 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#> 
#> $hierarch_single_manhattan_dtaEIdtaMEI$internal_metrics
#> $hierarch_single_manhattan_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.3540507
#> 
#> $hierarch_single_manhattan_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.1085543
#> 
#> $hierarch_single_manhattan_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.7129516
#> 
#> $hierarch_single_manhattan_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.1414405
#> 
#> 
#> $hierarch_single_manhattan_dtaEIdtaMEI$time
#> [1] 0.0004062653
#> 
#> 
#> $hierarch_single_manhattan_dtaHIdtaMHI
#> $hierarch_single_manhattan_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   1   2   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#> 
#> $hierarch_single_manhattan_dtaHIdtaMHI$internal_metrics
#> $hierarch_single_manhattan_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.4286669
#> 
#> $hierarch_single_manhattan_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.108974
#> 
#> $hierarch_single_manhattan_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.6451731
#> 
#> $hierarch_single_manhattan_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.1414405
#> 
#> 
#> $hierarch_single_manhattan_dtaHIdtaMHI$time
#> [1] 0.0004019737
#> 
#> 
#> $hierarch_complete_manhattan_dtaEIdtaMEI
#> $hierarch_complete_manhattan_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   2   1   1   1   2   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   2   1   1   1   1   2   2   2   1   1   1   2   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   2   2   2   1   1   1   2   1   2   2   2   2   2   2   2   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   1   2   1   2   1   2   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_complete_manhattan_dtaEIdtaMEI$internal_metrics
#> $hierarch_complete_manhattan_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.6494242
#> 
#> $hierarch_complete_manhattan_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.04254882
#> 
#> $hierarch_complete_manhattan_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.5421163
#> 
#> $hierarch_complete_manhattan_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.9814539
#> 
#> 
#> $hierarch_complete_manhattan_dtaEIdtaMEI$time
#> [1] 0.0004191399
#> 
#> 
#> $hierarch_complete_manhattan_dtaHIdtaMHI
#> $hierarch_complete_manhattan_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   1   1   1   1   1   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   2   2   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   2   1   1   1   2   2   2   2   1   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   1   2   2   1   1   2   1   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   1   1   2   1   2   1   1   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_complete_manhattan_dtaHIdtaMHI$internal_metrics
#> $hierarch_complete_manhattan_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.6103852
#> 
#> $hierarch_complete_manhattan_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.08346508
#> 
#> $hierarch_complete_manhattan_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.5562613
#> 
#> $hierarch_complete_manhattan_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.9765005
#> 
#> 
#> $hierarch_complete_manhattan_dtaHIdtaMHI$time
#> [1] 0.0004146099
#> 
#> 
#> $hierarch_average_manhattan_dtaEIdtaMEI
#> $hierarch_average_manhattan_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   2   2   2   1   1   2   2   2   2   2   2   1   2   2   2   2   2   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   2   2   2   2   2   1   2   1   2   2   2   2   2   1   1   2   1   2   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   2   2   2   2   1   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#> 
#> $hierarch_average_manhattan_dtaEIdtaMEI$internal_metrics
#> $hierarch_average_manhattan_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.4889417
#> 
#> $hierarch_average_manhattan_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.06555398
#> 
#> $hierarch_average_manhattan_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.6325931
#> 
#> $hierarch_average_manhattan_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.5842388
#> 
#> 
#> $hierarch_average_manhattan_dtaEIdtaMEI$time
#> [1] 0.0004127026
#> 
#> 
#> $hierarch_average_manhattan_dtaHIdtaMHI
#> $hierarch_average_manhattan_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   1   1   1   1   1   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   2   2   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   2   1   1   1   2   2   2   2   1   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   1   2   2   1   1   2   1   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   1   1   2   1   2   1   1   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_average_manhattan_dtaHIdtaMHI$internal_metrics
#> $hierarch_average_manhattan_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.6103852
#> 
#> $hierarch_average_manhattan_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.08346508
#> 
#> $hierarch_average_manhattan_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.5562613
#> 
#> $hierarch_average_manhattan_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.9765005
#> 
#> 
#> $hierarch_average_manhattan_dtaHIdtaMHI$time
#> [1] 0.0004239082
#> 
#> 
#> $hierarch_centroid_manhattan_dtaEIdtaMEI
#> $hierarch_centroid_manhattan_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   2   2   2   1   1   2   2   2   2   2   2   1   2   2   2   2   2   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   2   2   2   2   1   2   1   2   2   2   2   2   1   1   2   1   2   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   2   2   2   2   1   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2   2 
#> 
#> $hierarch_centroid_manhattan_dtaEIdtaMEI$internal_metrics
#> $hierarch_centroid_manhattan_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.5057735
#> 
#> $hierarch_centroid_manhattan_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.07548191
#> 
#> $hierarch_centroid_manhattan_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.6235726
#> 
#> $hierarch_centroid_manhattan_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.6098403
#> 
#> 
#> $hierarch_centroid_manhattan_dtaEIdtaMEI$time
#> [1] 0.0004127026
#> 
#> 
#> $hierarch_centroid_manhattan_dtaHIdtaMHI
#> $hierarch_centroid_manhattan_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   1   1   1   1   2   2   1   2   1   1   1   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   1   2   2   1   1   2   1   1   2   1   1   1   1   1   2   1   2   1   1   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   1   1   2   1   1   1   1   1   2   1   2   1   1   1   2   1   2 
#> 
#> $hierarch_centroid_manhattan_dtaHIdtaMHI$internal_metrics
#> $hierarch_centroid_manhattan_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.5274109
#> 
#> $hierarch_centroid_manhattan_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.06584967
#> 
#> $hierarch_centroid_manhattan_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.5985458
#> 
#> $hierarch_centroid_manhattan_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.7219281
#> 
#> 
#> $hierarch_centroid_manhattan_dtaHIdtaMHI$time
#> [1] 0.0004141331
#> 
#> 
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   2   1   1   1   2   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   2   1   2   1   1   2   2   2   1   1   1   2   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   2   2   2   1   1   1   2   1   2   2   2   2   2   2   2   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   2   2   2   2   2   2   2   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   2   1   2   1   2   1   2   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI$internal_metrics
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI$internal_metrics$davies_bouldin
#> [1] 0.6482362
#> 
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI$internal_metrics$dunn
#> [1] 0.05440341
#> 
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI$internal_metrics$silhouette
#> [1] 0.5447726
#> 
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI$internal_metrics$infomax
#> [1] 0.9765005
#> 
#> 
#> $hierarch_ward.D2_manhattan_dtaEIdtaMEI$time
#> [1] 0.0004115105
#> 
#> 
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI$cluster
#>   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18  19  20 
#>   1   2   1   1   1   1   1   1   1   2   1   2   2   1   1   2   1   1   1   1 
#>  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36  37  38  39  40 
#>   1   1   1   1   1   1   1   1   1   2   2   1   1   1   1   1   1   1   1   1 
#>  41  42  43  44  45  46  47  48  49  50  51  52  53  54  55  56  57  58  59  60 
#>   1   1   1   1   1   1   1   2   1   1   1   2   2   2   2   1   1   2   2   2 
#>  61  62  63  64  65  66  67  68  69  70  71  72  73  74  75  76  77  78  79  80 
#>   2   2   2   1   2   2   1   2   2   1   1   2   1   2   2   1   2   1   2   2 
#>  81  82  83  84  85  86  87  88  89  90  91  92  93  94  95  96  97  98  99 100 
#>   2   2   2   1   1   2   1   2   1   1   2   2   2   2   1   2   2   2   1   2 
#> 
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI$internal_metrics
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI$internal_metrics$davies_bouldin
#> [1] 0.6103852
#> 
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI$internal_metrics$dunn
#> [1] 0.08346508
#> 
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI$internal_metrics$silhouette
#> [1] 0.5562613
#> 
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI$internal_metrics$infomax
#> [1] 0.9765005
#> 
#> 
#> $hierarch_ward.D2_manhattan_dtaHIdtaMHI$time
#> [1] 0.0004093647
#> 
#>